requirements.txt
# Interface Web
streamlit

# Google Gemini API
google-generativeai

# Manipulação de PDF e OCR (Visão)
PyPDF2
pdf2image
pytesseract

# LangChain e Integrações
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters

# Banco de Dados Vetorial e Embeddings (HuggingFace)
faiss-cpu
sentence-transformers

# Variáveis de Ambiente
python-dotenv